chore: Extract SpecModel from Processor
No logical changes were made besides adapting the code to the new
interface.
